LayoutTests/fast/regions/offsetLeft-offsetTop-in-multiple-regions-expected.txt

Test offsetLeft and offsetTop in a named flow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS divText1.offsetParent is document.body
PASS divText1.offsetLeft is 100
PASS divText1.offsetTop is 450
PASS divText2_1.offsetParent is document.body
PASS divText2_1.offsetLeft is 500
PASS divText2_1.offsetTop is 150
PASS divText2_2.offsetParent is document.body
PASS divText2_2.offsetLeft is 450
PASS divText2_2.offsetTop is 350
PASS divText2_3.offsetParent is document.body
PASS divText2_3.offsetLeft is 600
PASS divText2_3.offsetTop is 500
PASS image.offsetParent is document.body
PASS image.offsetLeft is 650
PASS image.offsetTop is 620
PASS successfullyParsed is true

TEST COMPLETE
